feat: update Reteno SDK version to 3.0.0 and refactor roadmap, API specifications, and listener methods

+++ android/src/main/java/com/retenosdk/RetenoSdkModule.java
@@ -95,7 +95,7 @@ public class RetenoSdkModule extends ReactContextBaseJavaModule {
   }
 
   @ReactMethod
-  public void removeListeners(Integer count) {
+  public void removeListeners(double count) {
     // Required by NativeEventEmitter. Individual listener cleanup is handled on the JS side.
   }
